src/requirements.txt
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
psycopg2-binary==2.9.5
requests==2.28.2
django-environ==0.9.0
drf-spectacular==0.25.1
toml==0.10.2
model-bakery==1.10.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
pytest==7.2.1
black==23.1.0
